+++ R/aml_test_model.R
@@ -131,7 +131,7 @@ aml_test_model <- function(symbol, num_bars, timeframe, path_model, path_data,
     utils::head(num_bars)
 
   #dataset with tick data
-  z <- readr::read_csv(path_tick, col_names = FALSE) %>%
+  z <- readr::read_csv(path_tick, col_names = FALSE, col_types = readr::cols()) %>%
     #filter line with a symbol we need
     dplyr::filter(X1 == symbol) %$%
     #value z will contain tick value for this symbol