+++ NEWS.md
@@ -2,8 +2,22 @@
 
 ## Planned Changes
 
-* cleanup of deprecated functions
 * setup github actions
+* add fail safe for function input parameters
+
+# lazytrade 0.5.2
+
+# Version 0.5.2
+
+## Changes
+
+* add second parameter to simulation function `aml_simulation`
+* option to use full columns for model training when selecting 0 as a parameter `num_cols_used`
+* add suppress messages option during `readr::read_csv()` function calls
+* fail safe in `aml_collect_data` function will delete already recorded rds file if it has different amount of columns
+* add new function `util_find_pid` to find the PIDs of the terminal.exe application
+* function `mt_stat_transf` is now using a rule to assign 3 market type classes
+* rewrite function `mt_make_model` with the same philosophy as in `aml_make_model`
 
 # lazytrade 0.5.1